requirements.txt
# Core dependencies for AWS ETL Pipeline
apache-airflow[aws]==2.7.3
apache-airflow-providers-amazon==8.0.0
boto3==1.28.57
botocore==1.31.57

# Data processing
pandas==2.0.3
pyspark==3.4.1
awsglue==1.0.0

# dbt dependencies
dbt-core==1.6.0
dbt-athena==1.6.0

# Testing and validation
pytest==7.4.0
pytest-cov==4.1.0

# Monitoring and logging
structlog==23.1.0

# Environment management
python-dotenv==1.0.0

# Web framework for monitoring dashboard
streamlit==1.28.1
plotly==5.15.0
